requirements.txt
# Core Web App
streamlit>=1.33.0
altair>=5.0.0

# Data Processing
pandas>=2.2.0
numpy>=1.26.0
pyarrow>=15.0.0

# SQL & Database
SQLAlchemy>=2.0.0
duckdb>=0.10.2
sqlite-utils>=3.36.0
sqlparse>=0.4.4

# ML & Preprocessing
scikit-learn>=1.4.0
scipy>=1.12.0
joblib>=1.3.2
shap>=0.45.0
numba>=0.59.0
llvmlite>=0.42.0

# Vector Store + Embeddings (for RAG)
faiss-cpu>=1.8.0
sentence-transformers>=2.6.0
tqdm>=4.66.0

# LLM Clients
openai>=1.14.0
anthropic>=0.18.1
transformers>=4.39.0
huggingface-hub>=0.21.3

# LangChain
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-google-genai
langchain-anthropic
langchain-community>=0.0.20
langchain-text-splitters>=0.0.1
langgraph>=0.0.20

# Optional: DeepSeek, local models, Ollama
httpx>=0.27.0
requests>=2.31.0
kaggle>=1.6.0
PyGithub>=2.1.0

# PDF Export
reportlab>=4.0.9
fpdf2>=2.7.8

# File Parsing
PyPDF2>=3.0.0
python-docx>=1.1.0

# Visualization & EDA
streamlit-agraph>=0.0.45
graphviz>=0.20.1
pyvis>=0.3.2
matplotlib>=3.8.2
seaborn>=0.13.2

# JSON / YAML / Parsing
pydantic>=2.6.0
pydantic-settings>=2.2.0
PyYAML>=6.0.1
vega_datasets>=0.9.0

# Utilities
python-dateutil>=2.9.0
python-dotenv>=1.0.0
regex>=2023.12.25
# For downloading, file paths, compatibility
pathlib>=1.0.1
typing_extensions>=4.10.0

# FastAPI/Server backend
fastapi>=0.110.0
uvicorn>=0.29.0
python-multipart>=0.0.9
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0

# Logging
loguru>=0.7.2

# For PK/FK detection & ERD rendering
networkx>=3.2.1

# For RAG “context inspector”
markdown>=3.6
pygments>=2.17.2
neo4j>=5.0.0
fpdf==1.7.2
moviepy>=1.0.3 # For audio extraction from video
python-pptx
beautifulsoup4
python-docx>=0.6.21 # For PPT generation
